+++ tests/test_client.py
@@ -134,15 +134,15 @@ class TestGetwaveformSpanClass:
         with pytest.raises(ValueError):
             client.get_waveform('0101', starttime, 400000)
 
-    def test_get_waveform_subrequests_larger_than_120(self, client):
-        starttime = datetime(2005, 1, 1, 0, 0)
-        with pytest.raises(ValueError):
-            client.get_waveform('0101', starttime, 1440)
-
-    def test_get_waveform_wrong_max_span(self, client):
-        starttime = datetime(2005, 1, 1, 0, 0)
-        with pytest.raises(ValueError):
-            client.get_waveform('0101', starttime, 10, max_span=65)
+    def test_get_waveform_larger_max_span(self, client):
+        starttime = datetime(2010, 1, 1, 0, 0)
+        data, ctable = client.get_waveform('0101', starttime, 10, max_span=65)
+        assert data == '0101_201001010000_10.cnt'
+        assert os.path.exists(data)
+        os.remove(data)
+        assert ctable == '0101_20100101.ch'
+        assert os.path.exists(ctable)
+        os.remove(ctable)
 
     def test_get_waveform_wrong_starttime(self, client):
         starttime = datetime(2001, 1, 1, 0, 0)
@@ -196,11 +196,11 @@ class TestClientOthersClass:
             client._parse_code('01013')
 
     def test_get_allowed_span(self, client):
-        assert client.get_allowed_span('0401') == 60
+        assert client._get_allowed_span('0401') == 60
         client.select_stations('0101')
-        assert client.get_allowed_span('0101') == 5
+        assert client._get_allowed_span('0101') == 5
         client.select_stations('0101', ['N.AAKH', 'N.ABNH'])
-        assert client.get_allowed_span('0101') == 60
+        assert client._get_allowed_span('0101') == 60
 
     def test_info(self, client):
         client.info()